refactor: util functions for fetching book description and rendering book search component
