+++ client/components/UserBooks.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {fetchUserBooks, postUserBook, deleteUserBook} from '../store'
-import BookSearch from './BookSearch'
+import {makeBookObject, renderBookSearch} from '../utils'
 
 // Material UI
 import {withStyles} from '@material-ui/core/styles'
@@ -36,23 +36,9 @@ class UserBooks extends Component {
     this.setState({[`${type}Results`]: results})
   }
 
-  handleAddBook(e, bookResult, type) {
+  async handleAddBook(e, bookResult, type) {
     e.preventDefault()
-    // TODO: fetch book description and save to db
-    const newBook = {
-      author: bookResult.best_book.author,
-      goodReadsId: bookResult.best_book.id,
-      title: bookResult.best_book.title,
-      imageUrl: bookResult.best_book.image_url,
-      smallImageUrl: bookResult.best_book.small_image_url,
-      pubDate:
-        bookResult.original_publication_month +
-        '-' +
-        bookResult.original_publication_day +
-        '-' +
-        bookResult.original_publication_year,
-      rating: Math.round(bookResult.average_rating * 100)
-    }
+    const newBook = await makeBookObject(bookResult)
     this.props.postUserBook(newBook, type)
     this.setState({
       nowResults: [],
@@ -75,13 +61,7 @@ class UserBooks extends Component {
             : type === 'future' ? `Books I Want To Read` : `Books I've Read`}
         </Typography>
 
-        <BookSearch
-          bookList={books}
-          results={this.state[`${type}Results`]}
-          setResults={results => this.setResults(results, type)}
-          addBook={(e, book) => this.handleAddBook(e, book, type)}
-          removeBook={(e, idx, bookId) => this.handleRemoveBook(e, idx, bookId)}
-        />
+        {renderBookSearch(books, type, this)}
       </div>
     )
   }
@@ -100,11 +80,11 @@ class UserBooks extends Component {
         </Typography>
         <div>
           <Divider />
-          {this.renderBookSection(currentBooks, 'now', classes)}
+          {this.renderBookSection(currentBooks, 'now', this, classes)}
           <Divider />
-          {this.renderBookSection(futureBooks, 'future', classes)}
+          {this.renderBookSection(futureBooks, 'future', this, classes)}
           <Divider />
-          {this.renderBookSection(pastBooks, 'past', classes)}
+          {this.renderBookSection(pastBooks, 'past', this, classes)}
         </div>
       </div>
     )
